# docs/phase2/29_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: FlexSpace API
  version: "1.0.0"
  description: >
    FlexSpace Flexible Workspace Marketplace — Phase 2 API contract.
    Implements the decisions in 11_API_CONTRACTS.md. REST/JSON, versioned base path,
    JWT bearer auth, consistent error envelope, idempotency keys on mutating
    financial/booking endpoints. This is a contract, not an implementation —
    no server code exists yet.
servers:
  - url: https://api.flexspace.az/api/v1
    description: Production (placeholder domain)
  - url: https://staging-api.flexspace.az/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /payments:
    post:
      tags: [Payments]
      summary: Create a hosted checkout session (13_PAYMENT_ARCHITECTURE.md §13.3)
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, provider]
              properties:
                bookingId: { type: string, format: uuid }
                provider: { type: string, enum: [EPOINT, PAYRIFF] }
      responses:
        "200":
          description: Redirect URL to hosted checkout
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkoutUrl: { type: string, format: uri }
                  paymentId: { type: string, format: uuid }
  /payments/webhook/{provider}:
    post:
      tags: [Payments]
      summary: Provider webhook receiver (signature-verified, idempotent — 18_SECURITY.md §18.3)
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [epoint, payriff] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: "Raw provider-specific payload" }
      responses:
        "200": { description: Acknowledged }
        "400": { description: "Invalid signature — rejected, no state change" }
  /refunds:
    post:
      tags: [Payments]
      summary: Request a refund (computed per cancellation policy — 05_USER_FLOWS.md §5.6)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, reason]
              properties:
                bookingId: { type: string, format: uuid }
                reason: { type: string }
      responses:
        "202":
          description: Refund requested
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }

  # ------------------------------------------------------------------------
  # CUSTOMER ACCOUNT
  # ------------------------------------------------------------------------
...
  /provider/locations:
    get:
      tags: [Provider]
      summary: List my locations
      responses:
        "200":
          description: Locations
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Location" } }
    post:
      tags: [Provider]
      summary: Create a location (requires verification_status != REJECTED — 25_PROVIDER_ARCHITECTURE.md)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocationInput" }
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
  /provider/locations/{locationId}:
    patch:
      tags: [Provider]
      summary: Update a location
      parameters:
        - { name: locationId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocationInput" }
      responses: { "200": { description: Updated } }
...
  /admin/refunds/{refundId}/approve:
    post:
      tags: [Admin]
      summary: Approve a refund above the Support/Ops authorized limit (18_SECURITY.md §18.2)
      parameters:
        - { name: refundId, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: Refund approved } }
...
  /r/{code}:
    get:
      tags: [Partner]
      summary: Public referral tracking redirect (31_PARTNER_REFERRAL_ARCHITECTURE.md §31.4)
      description: >
        Looks up the ACTIVE ReferralCampaign by code, records a ReferralClick
        server-side, sets an httpOnly/Secure/SameSite=Lax cookie carrying only
        an opaque attribution_token, and redirects. No request/response body —
        the cookie is the only client-visible artifact, and it proves nothing
        by itself (the authoritative record is the server-side ReferralClick row).
      security: []
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        "302":
          description: Redirect to the marketplace homepage or landing path; Set-Cookie carries the attribution token
        "404":
          description: Unknown or inactive campaign code — redirects to the homepage without setting a cookie
  /admin/partners:
    get:
      tags: [Admin, Partner]
      summary: List/filter partners
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [PENDING, ACTIVE, SUSPENDED] } }
        - { name: type, in: query, schema: { type: string, enum: [AFFILIATE, AGENCY, CORPORATE, COMMUNITY, STRATEGIC_PARTNER] } }
      responses:
        "200":
          description: Partners
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Partner" } }
    post:
      tags: [Admin, Partner]
      summary: Create partner (admin-provisioned — no self-signup in V1, §31.7)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerCreateRequest" }
      responses:
        "201":
          description: Partner created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Partner" }
  /admin/partners/{partnerId}:
    patch:
      tags: [Admin, Partner]
      summary: Update a partner, including status transitions (PENDING/ACTIVE/SUSPENDED)
      parameters:
        - { name: partnerId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerUpdateRequest" }
      responses: { "200": { description: Partner updated } }
  /admin/partners/{partnerId}/campaigns:
    get:
      tags: [Admin, Partner]
      summary: List a partner's referral campaigns
      parameters:
        - { name: partnerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ReferralCampaign" } }
    post:
      tags: [Admin, Partner]
      summary: Create a referral campaign (trackable code/link) for a partner
      parameters:
        - { name: partnerId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReferralCampaignCreateRequest" }
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferralCampaign" }
  /admin/partners/{partnerId}/analytics:
    get:
      tags: [Admin, Partner]
      summary: Clicks, attributed bookings, commission earned/paid (§31.7 — admin-facing only in V1)
      parameters:
        - { name: partnerId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Partner analytics summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerAnalytics" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: 11_API_CONTRACTS.md §11.2 — required in practice for booking/payment mutations
...
  schemas:
...
    Money:
      type: object
      properties:
        amount: { type: integer, description: "minor units (qəpik)" }
        currency: { type: string, example: "AZN" }
...
    Refund:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        amount: { $ref: "#/components/schemas/Money" }
        status: { type: string, enum: [REQUESTED, APPROVED, PROCESSING, COMPLETED, REJECTED] }
...
    Location:
      type: object
      properties:
        id: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        name: { type: string }
        addressLine: { type: string }
        city: { type: string }
        district: { type: string, nullable: true }
        countryCode: { type: string, example: "AZ" }
        lat: { type: number }
        lng: { type: number }
        timezone: { type: string, example: "Asia/Baku" }
        openingHours: { type: object, nullable: true }
    LocationInput:
      type: object
      required: [name, addressLine, city, lat, lng]
      properties:
        name: { type: string }
        addressLine: { type: string }
        city: { type: string }
        district: { type: string }
        countryCode: { type: string, default: "AZ" }
        lat: { type: number }
        lng: { type: number }
        timezone: { type: string, default: "Asia/Baku" }
        openingHours: { type: object }
...
    Partner:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string, enum: [AFFILIATE, AGENCY, CORPORATE, COMMUNITY, STRATEGIC_PARTNER] }
        status: { type: string, enum: [PENDING, ACTIVE, SUSPENDED] }
        contactEmail: { type: string, format: email, nullable: true }
        contactPhone: { type: string, nullable: true }
        defaultCommissionType: { type: string, enum: [PERCENTAGE_OF_PLATFORM_FEE, FIXED_PER_BOOKING] }
        defaultCommissionValue: { type: integer, description: "Minor units if FIXED_PER_BOOKING, basis points if PERCENTAGE_OF_PLATFORM_FEE" }
        createdAt: { type: string, format: date-time }
    PartnerCreateRequest:
      type: object
      required: [name, type, defaultCommissionType, defaultCommissionValue]
      properties:
        name: { type: string }
        type: { type: string, enum: [AFFILIATE, AGENCY, CORPORATE, COMMUNITY, STRATEGIC_PARTNER] }
        contactEmail: { type: string, format: email }
        contactPhone: { type: string }
        defaultCommissionType: { type: string, enum: [PERCENTAGE_OF_PLATFORM_FEE, FIXED_PER_BOOKING] }
        defaultCommissionValue: { type: integer }
    PartnerUpdateRequest:
      type: object
      properties:
        name: { type: string }
        status: { type: string, enum: [PENDING, ACTIVE, SUSPENDED] }
        contactEmail: { type: string, format: email }
        contactPhone: { type: string }
        defaultCommissionType: { type: string, enum: [PERCENTAGE_OF_PLATFORM_FEE, FIXED_PER_BOOKING] }
        defaultCommissionValue: { type: integer }
    ReferralCampaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        partnerId: { type: string, format: uuid }
        name: { type: string }
        code: { type: string, example: "BLOGGER123" }
        status: { type: string, enum: [ACTIVE, PAUSED, ENDED] }
        attributionWindowDays: { type: integer, example: 30 }
        commissionTypeOverride: { type: string, enum: [PERCENTAGE_OF_PLATFORM_FEE, FIXED_PER_BOOKING], nullable: true }
        commissionValueOverride: { type: integer, nullable: true }
        startsAt: { type: string, format: date-time, nullable: true }
        endsAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ReferralCampaignCreateRequest:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string, description: "Unique, human-facing (e.g. a blogger's handle)" }
        attributionWindowDays: { type: integer, default: 30 }
        commissionTypeOverride: { type: string, enum: [PERCENTAGE_OF_PLATFORM_FEE, FIXED_PER_BOOKING] }
        commissionValueOverride: { type: integer }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
    PartnerAnalytics:
      type: object
      properties:
        partnerId: { type: string, format: uuid }
        totalClicks: { type: integer }
        attributedBookings: { type: integer }
        commissionEarned: { $ref: "#/components/schemas/Money" }
        commissionPaid: { $ref: "#/components/schemas/Money" }
        commissionPending: { $ref: "#/components/schemas/Money" }
        byCampaign:
          type: array
          items:
            type: object
            properties:
              campaignId: { type: string, format: uuid }
              code: { type: string }
              clicks: { type: integer }
              attributedBookings: { type: integer }
              commissionEarned: { $ref: "#/components/schemas/Money" }